Rename using capital first letter for consistency
