+++ src/models/Ticket.ts
